myfirst-agent/pyproject.toml
[project]
name = "myfirst-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Muhammad Asif", email = "programmerasif@outlook.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]

[project.scripts]
myfirst-agent = "myfirst_agent:main"
myfirst-agent1 = "myfirst_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
